separated importdlg bit of a rigged solution, but hopefully encourages more code cleanup
